+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### API Changes
 
 ### Enhancements
-* Client side: wc/dom/getLabelsForElement loads <label> for elements that define aria-labelledby attribute #1401
+* Client side: wc/dom/getLabelsForElement an enhancement to load elements defined by 'aria-labelledby' attribute #1401.
 
 ### Bug Fixes